+++ src/pages/app/Vendors/ProposalOutside/index.vue
@@ -70,15 +70,15 @@ export default {
     await this.$store.dispatch('common/getTaxes');
 
     // handling uploading photo backhand process
-      this.$root.$on("update-inspirational-photo", async ({ file, index, link, fileName}) => {
+      this.$root.$on("update-inspirational-photo", async ({ file, index, link}) => {
           const currentPhoto = this.inspirationalPhotos[index];
 
-          const url = await S3Service.fileUpload(file, fileName, link)
+          let url = await S3Service.fileUpload(file, `photo-${index}`, link)
           this.$store.commit("proposalForNonMaryoku/setInspirationalPhoto", { index, photo: { ...currentPhoto, url }});
 
       });
       this.$root.$on("remove-inspirational-photo", async (index) => {
-          await S3Service.deleteFile(this.inspirationalPhotos[index].url);
+          if ( this.version !== -1 ) await S3Service.deleteFile(this.inspirationalPhotos[index].url);
           this.$store.commit("proposalForNonMaryoku/setInspirationalPhoto", { index, photo: null });
       })
   },
@@ -135,6 +135,9 @@ export default {
     inspirationalPhotos(){
       return this.$store.state.proposalForNonMaryoku.inspirationalPhotos;
     },
+    version(){
+      return this.$store.state.proposalForNonMaryoku.currentVersion;
+    }
   },
   watch: {
       taxes(newVal){console.log('taxes', newVal)}
